requirements.txt
httplib2==0.9.2
oauth2==1.9.0.post1
pyaml==16.9.0
python-dateutil==2.5.3
PyYAML==3.12
six==1.10.0
SQLAlchemy==1.1.2
wheel==0.24.0
